navbar v1.2 [completed]
