+++ package.json
@@ -17,6 +17,7 @@
     "next": "12.3.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-icons": "^4.4.0",
     "styled-components": "^5.3.6"
   },
   "devDependencies": {
